+++ lib/packages/http/src/httpAdapterAxios.ts
@@ -8,7 +8,7 @@ const DefaultAxiosOptions = {};
 /**
  * Http Implementation of [[Http]] using https://github.com/axios/axios
  *
- * You can use [[HttpClient]] as alias
+ * Prefer [[HttpClientFactory.createHttpClient]] to create an instance
  *
  * @module http
  */