+++ web/angular-wallet/src/@fuse/components/shortcuts/shortcuts.component.ts
@@ -32,18 +32,8 @@ export class FuseShortcutsComponent implements OnInit, OnDestroy {
   @ViewChild('shortcuts', {static: true})
   shortcutsEl: ElementRef;
 
-  // Private
   private _unsubscribeAll: Subject<any>;
 
-  /**
-   * Constructor
-   *
-   * @param {Renderer2} _renderer
-   * @param {CookieService} _cookieService
-   * @param {FuseMatchMediaService} _fuseMatchMediaService
-   * @param {FuseNavigationService} _fuseNavigationService
-   * @param {ObservableMedia} _observableMedia
-   */
   constructor(
     private i18nService: I18nService,
     private _cookieService: CookieService,
@@ -53,22 +43,13 @@ export class FuseShortcutsComponent implements OnInit, OnDestroy {
     private _renderer: Renderer2,
     private accountService: AccountService
   ) {
-    // Set the defaults
     this.shortcutItems = [];
     this.searching = false;
     this.mobileShortcutsPanelActive = false;
 
-    // Set the private defaults
     this._unsubscribeAll = new Subject();
   }
 
-  // -----------------------------------------------------------------------------------------------------
-  // @ Lifecycle hooks
-  // -----------------------------------------------------------------------------------------------------
-
-  /**
-   * On init
-   */
   ngOnInit(): void {
     // Get the navigation items and flatten them
     this.filteredNavigationItems = this.navigationItems = this._fuseNavigationService.getFlatNavigation(this.navigation);
@@ -83,7 +64,6 @@ export class FuseShortcutsComponent implements OnInit, OnDestroy {
       this.updateShortcuts();
     });
 
-    // Subscribe to media changes
     this._fuseMatchMediaService.onMediaChange
       .pipe(takeUntil(this._unsubscribeAll))
       .subscribe(() => {
@@ -102,58 +82,50 @@ export class FuseShortcutsComponent implements OnInit, OnDestroy {
 
   private updateShortcuts(): void {
     this.shortcutItems = [];
+    const isOfflineAccount = this.currentAccount && this.currentAccount.type === 'offline';
+
     this.shortcutItems.push({
-      'title': this.i18nService.getTranslation('dashboard'),
-      'type': 'item',
-      'icon': 'dashboard',
-      'url': '/dashboard'
+      title: this.i18nService.getTranslation('dashboard'),
+      type: 'item',
+      icon: 'dashboard',
+      url: '/dashboard'
     });
-    if (this.currentAccount && this.currentAccount.type !== 'offline') {
+    if (!isOfflineAccount) {
       this.shortcutItems.push({
-        'title': this.i18nService.getTranslation('send_burst'),
-        'type': 'item',
-        'icon': 'vertical_align_top',
-        'url': '/send'
+        title: this.i18nService.getTranslation('send_burst'),
+        type: 'item',
+        icon: 'vertical_align_top',
+        url: '/send'
       });
     }
     this.shortcutItems.push({
-        'title': this.i18nService.getTranslation('request_burst'),
-        'type': 'item',
-        'icon': 'vertical_align_bottom',
-        'url': '/request'
-      },
-      {
-        'title': this.i18nService.getTranslation('messages'),
-        'type': 'item',
-        'icon': 'message',
-        'url': '/messages'
-      },
-      {
-        'title': this.i18nService.getTranslation('settings'),
-        'type': 'item',
-        'icon': 'settings',
-        'url': '/settings'
-      });
+      title: this.i18nService.getTranslation('request_burst'),
+      type: 'item',
+      icon: 'vertical_align_bottom',
+      url: '/request'
+    });
+    if (!isOfflineAccount) {
+      this.shortcutItems.push(
+        {
+          title: this.i18nService.getTranslation('messages'),
+          type: 'item',
+          icon: 'message',
+          url: '/messages'
+        });
+    }
+    this.shortcutItems.push({
+      title: this.i18nService.getTranslation('settings'),
+      type: 'item',
+      icon: 'settings',
+      url: '/settings'
+    });
   }
 
-  /**
-   * On destroy
-   */
   ngOnDestroy(): void {
-    // Unsubscribe from all subscriptions
     this._unsubscribeAll.next();
     this._unsubscribeAll.complete();
   }
 
-  // -----------------------------------------------------------------------------------------------------
-  // @ Public methods
-  // -----------------------------------------------------------------------------------------------------
-
-  /**
-   * Search
-   *
-   * @param event
-   */
   search(event): void {
     const value = event.target.value.toLowerCase();
 
@@ -171,64 +143,38 @@ export class FuseShortcutsComponent implements OnInit, OnDestroy {
     });
   }
 
-  /**
-   * Toggle shortcut
-   *
-   * @param event
-   * @param itemToToggle
-   */
   toggleShortcut(event, itemToToggle): void {
     event.stopPropagation();
 
     for (let i = 0; i < this.shortcutItems.length; i++) {
       if (this.shortcutItems[i].url === itemToToggle.url) {
         this.shortcutItems.splice(i, 1);
-
-        // Save to the cookies
         this._cookieService.set('FUSE2.shortcuts', JSON.stringify(this.shortcutItems));
-
         return;
       }
     }
 
     this.shortcutItems.push(itemToToggle);
-
-    // Save to the cookies
     this._cookieService.set('FUSE2.shortcuts', JSON.stringify(this.shortcutItems));
   }
 
-  /**
-   * Is in shortcuts?
-   *
-   * @param navigationItem
-   * @returns {any}
-   */
   isInShortcuts(navigationItem): any {
     return this.shortcutItems.find(item => {
       return item.url === navigationItem.url;
     });
   }
 
-  /**
-   * On menu open
-   */
   onMenuOpen(): void {
     setTimeout(() => {
       this.searchInputField.nativeElement.focus();
     });
   }
 
-  /**
-   * Show mobile shortcuts
-   */
   showMobileShortcutsPanel(): void {
     this.mobileShortcutsPanelActive = true;
     this._renderer.addClass(this.shortcutsEl.nativeElement, 'show-mobile-panel');
   }
 
-  /**
-   * Hide mobile shortcuts
-   */
   hideMobileShortcutsPanel(): void {
     this.mobileShortcutsPanelActive = false;
     this._renderer.removeClass(this.shortcutsEl.nativeElement, 'show-mobile-panel');