+++ web/angular-wallet/src/app/app.component.ts
@@ -4,6 +4,7 @@ import {Platform} from '@angular/cdk/platform';
 import {AccountService} from './setup/account/account.service';
 import {StoreService} from './store/store.service';
 import {Account, BlockchainStatus} from '@burstjs/core';
+import {parseDeeplink} from '@burstjs/util';
 import {NotifierService} from 'angular-notifier';
 import {NetworkService} from './network/network.service';
 import {UtilService} from './util.service';
@@ -19,7 +20,7 @@ import {version} from '../../package.json';
 import {AppService} from './app.service';
 import {UnsubscribeOnDestroy} from './util/UnsubscribeOnDestroy';
 import {takeUntil} from 'rxjs/operators';
-import { Router, DefaultUrlSerializer, UrlSegmentGroup, UrlSegment, PRIMARY_OUTLET } from '@angular/router';
+import {Router, DefaultUrlSerializer, UrlSegmentGroup, UrlSegment, PRIMARY_OUTLET} from '@angular/router';
 
 @Component({
   selector: 'app',
@@ -99,22 +100,48 @@ export class AppComponent extends UnsubscribeOnDestroy implements OnInit, OnDest
     clearInterval(this.blockchainStatusInterval);
   }
 
-
   private initDeepLinkHandler(): void {
-    this.appService.onIpcMessage('deep-link-clicked', (url) => {
-      const parsedUrl = this.urlSerializer.parse(url.replace('burst://', ''));
-      const g: UrlSegmentGroup = parsedUrl.root.children[PRIMARY_OUTLET];
-      const s: UrlSegment[] = g.segments;
-      this.router.navigate([s[0].path.replace('requestBurst', 'send')], {
-        queryParams: parsedUrl.queryParams,
-        queryParamsHandling: 'merge'
-      });
+    this.appService.onIpcMessage('deep-link-clicked', async (url) => {
+      const deeplinkUrl = new URL(url);
+      if (deeplinkUrl.pathname === '//v1') {
+        await this.routeCIP22Deeplink(url);
+      } else {
+        await this.routeLegacyDeeplink(url);
+      }
       // fixes an issue with the view not rendering
       setTimeout(() => {
         this.applicationRef.tick();
       }, 1000);
     });
   }
+
+  private async routeCIP22Deeplink(url: string): Promise<void> {
+    const parts = parseDeeplink(url);
+    let route = '';
+    // do the routing here
+    switch (parts.action) {
+      case 'send-amount':
+      case 'pay':
+        route = 'send';
+        break;
+      default:
+        this.notifierService.notify('warning', `Unknown deep link action: ${parts.action}`);
+    }
+
+    const {payload} = parts;
+    await this.router.navigate([route], {queryParams: {cip22: true, payload}});
+  }
+
+  private async routeLegacyDeeplink(url: string): Promise<void> {
+    const parsedUrl = this.urlSerializer.parse(url.replace('burst://', ''));
+    const g: UrlSegmentGroup = parsedUrl.root.children[PRIMARY_OUTLET];
+    const s: UrlSegment[] = g.segments;
+    await this.router.navigate([s[0].path.replace('requestBurst', 'send')], {
+      queryParams: parsedUrl.queryParams,
+      queryParamsHandling: 'merge'
+    });
+  }
+
 
   private initRouteToHandler(): void {
     this.appService.onIpcMessage('route-to', (url) => {
@@ -183,7 +210,7 @@ export class AppComponent extends UnsubscribeOnDestroy implements OnInit, OnDest
         await this.accountService.synchronizeAccount(this.selectedAccount).catch(() => {
         });
         this.accountService.setCurrentAccount(this.selectedAccount);
-      // hit this call again every 1 sec if the blockchain is being downloaded
+        // hit this call again every 1 sec if the blockchain is being downloaded
       } else if (this.downloadingBlockchain) {
         setTimeout(this.checkBlockchainStatus.bind(this), 1000);
       }
@@ -219,14 +246,16 @@ export class AppComponent extends UnsubscribeOnDestroy implements OnInit, OnDest
 
 
   private openNewVersionDialog(): MatDialogRef<NewVersionDialogComponent> {
-     return this.newVersionDialog.open(NewVersionDialogComponent, {
+    return this.newVersionDialog.open(NewVersionDialogComponent, {
       data: this.updateInfo
     });
   }
 
   onClickedNewVersion(): void {
 
-    if (this.isDownloadingUpdate) { return; }
+    if (this.isDownloadingUpdate) {
+      return;
+    }
 
     this.openNewVersionDialog()
       .afterClosed()