+++ web/angular-wallet/src/app/main/transactions/transaction-table/transaction-table.component.ts
@@ -15,11 +15,12 @@ import {
   TransactionMiningSubtype,
   TransactionType
 } from '@burstjs/core';
-import {convertBurstTimeToDate, convertNQTStringToNumber, BurstValue} from '@burstjs/util';
+import {convertBurstTimeToDate, convertNQTStringToNumber, BurstValue, BurstTime} from '@burstjs/util';
 import {UtilService} from 'app/util.service';
 import {takeUntil} from 'rxjs/operators';
 import {UnsubscribeOnDestroy} from '../../../util/UnsubscribeOnDestroy';
 import {StoreService} from '../../../store/store.service';
+import {formatDate} from '@angular/common';
 
 @Component({
   selector: 'app-transaction-table',
@@ -69,15 +70,15 @@ export class TransactionTableComponent extends UnsubscribeOnDestroy implements A
     return address && address === this.account.accountRS;
   }
 
-  public getAmount(transaction: Transaction): number {
+  public getAmount(transaction: Transaction): string {
 
     if (this.isOwnAccount(transaction.senderRS)) {
-      return -this.convertNQTStringToNumber(transaction.amountNQT);
+      return BurstValue.fromPlanck(transaction.amountNQT).multiply(-1).getBurst();
     }
 
     return this.isMultiOutPayment(transaction)
-      ? getRecipientsAmount(this.account.account, transaction)
-      : this.convertNQTStringToNumber(transaction.amountNQT);
+      ? getRecipientsAmount(this.account.account, transaction).toString(10)
+      : BurstValue.fromPlanck(transaction.amountNQT).getBurst();
   }
 
   public isAmountNegative(transaction: Transaction): boolean {
@@ -94,5 +95,31 @@ export class TransactionTableComponent extends UnsubscribeOnDestroy implements A
 
   getCommitmentAmount(transaction): string {
     return BurstValue.fromPlanck(transaction.attachment.amountNQT || '0').getBurst();
+  }
+
+  getDate(tx: Transaction): string {
+    const time = BurstTime.fromBurstTimestamp(tx.timestamp);
+    return formatDate(time.getDate(), 'short', this.locale);
+  }
+
+  getRowClass(row: Transaction): string {
+
+    const cx = className => !row.confirmations ? `${className} unconfirmed` : className;
+
+    if (
+      (!row.recipient && (row.type !== TransactionType.Payment)) ||
+      (row.recipient === this.account.account && row.sender === this.account.account)
+    ) {
+      return cx('self');
+    }
+
+    if (
+      (row.recipient === this.account.account && row.sender !== this.account.account) ||
+      (!row.recipient && (row.type === TransactionType.Payment)) //
+    ) {
+      return cx('incoming');
+    }
+
+    return cx('outgoing');
   }
 }